.history/backend/requirements_20250812232044.txt
# Django e DRF
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3
django-extensions==3.2.3

# Autenticação e Segurança
djangorestframework-simplejwt==5.3.0
django-allauth==0.57.0

# Banco de Dados
psycopg2-binary==2.9.7
dj-database-url==2.1.0

# Processamento de Excel
pandas==2.1.3
openpyxl==3.1.2
xlrd==2.0.1

# Cache e Filas
redis==5.0.1
celery==5.3.4
django-celery-beat==2.5.0

# Validação e Serialização
marshmallow==3.20.1
drf-yasg==1.21.7

# Utilitários
python-decouple==3.8
Pillow==10.1.0
reportlab==4.0.7

# Desenvolvimento
django-debug-toolbar==4.2.0
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0

# Produção
gunicorn==21.2.0
whitenoise==6.6.0 
